refactor: NavItem과 FooterItem 컴포넌트를 별도로 분리하여 구조 개선
