+++ src/components/Sidebar/Sidebar.tsx
@@ -1,6 +1,6 @@
-import { ReactNode, useMemo } from 'react';
+import { useMemo } from 'react';
 
-import { Link, useMatches } from '@tanstack/react-router';
+import { useMatches } from '@tanstack/react-router';
 
 import {
   CheckRightIcon,
@@ -16,45 +16,10 @@ import {
   UserIcon,
 } from '@/assets/icons';
 
+import { FooterItem } from './FooterItem';
+import { NavItem } from './NavItem';
 import { useSidebar } from './context';
 import * as styles from './sidebar.css';
-
-interface NavItemProps {
-  to: string;
-  icon: ReactNode;
-  label?: string;
-  isActive: boolean;
-  isOpen: boolean;
-}
-
-function NavItem({ to, icon, label, isActive, isOpen }: NavItemProps) {
-  return (
-    <Link to={to} className={styles.navItem({ active: isActive, isOpen })}>
-      <div className={styles.navItemIcon({ isOpen })}>{icon}</div>
-      {label && <span className={styles.navLabel}>{label}</span>}
-    </Link>
-  );
-}
-
-interface FooterItemProps {
-  to: string;
-  icon: ReactNode;
-  label?: string;
-  showArrow?: boolean;
-  isOpen: boolean;
-}
-
-function FooterItem({ to, icon, label, showArrow, isOpen }: FooterItemProps) {
-  return (
-    <Link to={to} className={styles.footerItem({ isOpen })}>
-      <div className={styles.footerItemIcon({ isOpen })}>
-        {icon}
-        {label && <span className={styles.footerLabel}>{label}</span>}
-      </div>
-      {showArrow && <CheckRightIcon className={styles.rightIcon} />}
-    </Link>
-  );
-}
 
 export function Sidebar() {
   const { isOpen, toggle, showToggle } = useSidebar();
@@ -110,8 +75,10 @@ export function Sidebar() {
   return (
     <aside className={styles.container({ isOpen })}>
       {showToggle && (
-        <button onClick={toggle} className={styles.toggleButton({ isOpen })}>
-          <CheckRightIcon className={styles.icon} />
+        <button onClick={toggle} className={styles.toggleButton}>
+          <CheckRightIcon
+            className={`${styles.rightIcon} ${isOpen ? styles.rotateIcon : ''}`}
+          />
         </button>
       )}
