+++ src/components/Sidebar/index.ts
@@ -1,2 +1,4 @@
 export * from './Sidebar';
 export * from './context';
+export * from './NavItem';
+export * from './FooterItem';
